requirements.txt
backcall==0.2.0
chumpy==0.70
cycler==0.10.0
dataclasses==0.7
decorator==4.4.2
distro==1.5.0
future==0.18.2
fvcore==0.1.2.post20201111
imageio==2.9.0
importlib-metadata==2.0.0
ipdb==0.13.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
jsonpickle==1.4.1
kiwisolver==1.3.1
lmdb==1.0.0
MarkupSafe==1.1.1
matplotlib==3.3.3
networkx==2.5
numpy==1.19.4
opencv-python==4.4.0.46
packaging==20.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pkg-resources==0.0.0
portalocker==2.0.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
-e git+https://github.com/ParthaEth/pytorch3d.git
pyvis==0.1.8.2
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-build==0.11.1
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tabulate==0.8.7
termcolor==1.1.0
tifffile==2020.9.3
torch==1.7.0
torchvision==0.8.1
tqdm==4.51.0
traitlets==4.3.3
typing-extensions==3.7.4.3
wcwidth==0.2.5
wrapt==1.12.1
yacs==0.1.8
zipp==3.4.0
smplx==0.1.26
torchgeometry==0.1.2
